+++ frontend/src/pages/index.ts
@@ -1,3 +1,4 @@
 export * from './bills';
+export * from './cashflow';
 export * from './home';
 export * from './income';